chore: handle cross-account trade filtering
