+++ src/state/apis/swappers/swappersApi.ts
@@ -7,6 +7,7 @@ import { getInputOutputRatioFromQuote } from 'state/apis/swappers/helpers/getInp
 import { getLifiTradeQuoteHelper } from 'state/apis/swappers/helpers/getLifiTradeQuoteApiHelper'
 import { getThorTradeQuoteHelper } from 'state/apis/swappers/helpers/getThorTradeQuoteApiHelper'
 import type { ApiQuote } from 'state/apis/swappers/types'
+import { isCrossAccountTradeSupported } from 'state/helpers'
 import type { ReduxState } from 'state/reducer'
 import type { FeatureFlags } from 'state/slices/preferencesSlice/preferencesSlice'
 import { selectFeatureFlags } from 'state/slices/selectors'
@@ -23,6 +24,8 @@ export const swappersApi = createApi({
     getTradeQuote: build.query<ApiQuote[], GetTradeQuoteInput>({
       queryFn: async (getTradeQuoteInput: GetTradeQuoteInput, { getState, dispatch }) => {
         const state = getState() as ReduxState
+        const { sendAddress, receiveAddress } = getTradeQuoteInput
+        const isCrossAccountTrade = sendAddress !== receiveAddress
         const { LifiSwap, ThorSwap }: FeatureFlags = selectFeatureFlags(state)
         const quotes: (Result<TradeQuote2, SwapErrorRight> & {
           swapperName: SwapperName
@@ -38,18 +41,25 @@ export const swappersApi = createApi({
             ...(await getThorTradeQuoteHelper(...quoteHelperArgs)),
             swapperName: SwapperName.Thorchain,
           })
-        const quotesWithInputOutputRatios = quotes.map(result => {
-          const quote = result && result.isOk() ? result.unwrap() : undefined
-          const error = result && result.isErr() ? result.unwrapErr() : undefined
-          const inputOutputRatio = quote
-            ? getInputOutputRatioFromQuote({
-                state,
-                quote,
-                swapperName: result.swapperName,
-              })
-            : -Infinity
-          return { quote, error, inputOutputRatio, swapperName: result.swapperName }
-        })
+        const quotesWithInputOutputRatios = quotes
+          .map(result => {
+            const quote = result && result.isOk() ? result.unwrap() : undefined
+            const error = result && result.isErr() ? result.unwrapErr() : undefined
+            const inputOutputRatio = quote
+              ? getInputOutputRatioFromQuote({
+                  state,
+                  quote,
+                  swapperName: result.swapperName,
+                })
+              : -Infinity
+            return { quote, error, inputOutputRatio, swapperName: result.swapperName }
+          })
+          .filter(result => {
+            const swapperSupportsCrossAccountTrade = isCrossAccountTradeSupported(
+              result.swapperName,
+            )
+            return !isCrossAccountTrade || swapperSupportsCrossAccountTrade
+          })
         // fixme: filter supports cross-account
         //   const isCrossAccountTrade = useMemo(
         //     () => sellAccountId !== buyAccountId,
