+++ src/state/slices/tradeQuoteSlice/helpers.ts
@@ -3,9 +3,7 @@ import { getDefaultSlippagePercentageForSwapper } from 'constants/constants'
 import type { BigNumber } from 'lib/bignumber/bignumber'
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
-import type { ProtocolFee, TradeQuote } from 'lib/swapper/api'
-import { SwapperName } from 'lib/swapper/api'
-import { assertUnreachable } from 'lib/utils'
+import type { ProtocolFee, SwapperName, TradeQuote } from 'lib/swapper/api'
 import { selectFeeAssetById } from 'state/slices/assetsSlice/selectors'
 import {
   selectCryptoMarketData,
@@ -172,22 +170,3 @@ export const getTotalProtocolFeeByAsset = (quote: TradeQuote): Record<AssetId, P
       acc,
     )
   }, {})
-
-export const isCrossAccountTradeSupported = (swapperName: SwapperName) => {
-  switch (swapperName) {
-    case SwapperName.Thorchain:
-    case SwapperName.Osmosis:
-      return true
-    // NOTE: Before enabling cross-account for LIFI and OneInch - we must pass the sending address
-    // to the swappers up so allowance checks work. They're currently using the receive address
-    // assuming it's the same address as the sending address.
-    case SwapperName.LIFI:
-    case SwapperName.OneInch:
-    case SwapperName.Zrx:
-    case SwapperName.CowSwap:
-    case SwapperName.Test:
-      return false
-    default:
-      assertUnreachable(swapperName)
-  }
-}
